requirements.txt
flask==3.0.3
werkzeug==3.0.3
Pillow==10.4.0
numpy==1.26.4
cryptography==46.0.3
